# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DesignGuard API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface ScanRequest {
  userId: string;
  url: string;
  screenshotBase64: string;
}

export type FixActionType = (typeof FixActionType)[keyof typeof FixActionType];

export const FixActionType = {
  css_injection: "css_injection",
  attribute_injection: "attribute_injection",
  devtools_command: "devtools_command",
} as const;

export interface FixAction {
  type: FixActionType;
  property?: string;
  oldValue?: string;
  newValue?: string;
  devtoolsCommand: string;
}

export type ViolationCategory =
  (typeof ViolationCategory)[keyof typeof ViolationCategory];

export const ViolationCategory = {
  COLOR_CONTRAST: "COLOR_CONTRAST",
  FONT_SIZE: "FONT_SIZE",
  MISSING_ALT_TEXT: "MISSING_ALT_TEXT",
  KEYBOARD_FOCUS: "KEYBOARD_FOCUS",
  BUTTON_SIZE: "BUTTON_SIZE",
  HEADING_STRUCTURE: "HEADING_STRUCTURE",
  LINK_TEXT: "LINK_TEXT",
  FORM_LABELS: "FORM_LABELS",
} as const;

export type ViolationSeverity =
  (typeof ViolationSeverity)[keyof typeof ViolationSeverity];

export const ViolationSeverity = {
  critical: "critical",
  serious: "serious",
  moderate: "moderate",
} as const;

export interface Violation {
  violationId: string;
  category: ViolationCategory;
  severity: ViolationSeverity;
  elementDescription: string;
  wcagCriterion: string;
  currentValue?: string;
  requiredValue?: string;
  fixDescription: string;
  fixAction?: FixAction;
}

export interface AppliedFix {
  violationId: string;
  fixApplied: boolean;
  devtoolsCommand: string;
  explanation: string;
  wcagCriterionMet: string;
  complianceImprovement: number;
}

export interface ScanResult {
  scanId: string;
  violationsFound: Violation[];
  fixesApplied: AppliedFix[];
  complianceScoreBefore: number;
  complianceScoreAfter: number;
  totalViolations: number;
  criticalCount: number;
  seriousCount: number;
  moderateCount: number;
  pageSummary: string;
}

export interface ApplyFixRequest {
  scanId: string;
  violationId: string;
}

export interface FixDetail {
  actionType: string;
  targetElement: string;
  fixCode: string;
  explanation: string;
}

export interface ScanSummary {
  scanId: string;
  url: string;
  createdAt: string;
  complianceScoreBefore: number;
  complianceScoreAfter: number;
  totalViolations: number;
  criticalCount: number;
  seriousCount: number;
  moderateCount: number;
  pageSummary: string;
}

export interface ScanHistoryResponse {
  scans: ScanSummary[];
}