# lib/api-zod/src/generated/types/appliedFix.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DesignGuard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppliedFix {
  violationId: string;
  fixApplied: boolean;
  devtoolsCommand: string;
  explanation: string;
  wcagCriterionMet: string;
  complianceImprovement: number;
}