# lib/api-zod/src/generated/types/fixAction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DesignGuard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FixActionType } from "./fixActionType";

export interface FixAction {
  type: FixActionType;
  property?: string;
  oldValue?: string;
  newValue?: string;
  devtoolsCommand: string;
}