# lib/api-zod/src/generated/types/fixDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DesignGuard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FixDetail {
  actionType: string;
  targetElement: string;
  fixCode: string;
  explanation: string;
}