# lib/api-zod/src/generated/types/scanHistoryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DesignGuard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanSummary } from "./scanSummary";

export interface ScanHistoryResponse {
  scans: ScanSummary[];
}