# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DesignGuard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppliedFix } from "./appliedFix";
import type { Violation } from "./violation";

export interface ScanResult {
  scanId: string;
  violationsFound: Violation[];
  fixesApplied: AppliedFix[];
  complianceScoreBefore: number;
  complianceScoreAfter: number;
  totalViolations: number;
  criticalCount: number;
  seriousCount: number;
  moderateCount: number;
  pageSummary: string;
}