# lib/api-zod/src/generated/types/scanSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DesignGuard API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScanSummary {
  scanId: string;
  url: string;
  createdAt: string;
  complianceScoreBefore: number;
  complianceScoreAfter: number;
  totalViolations: number;
  criticalCount: number;
  seriousCount: number;
  moderateCount: number;
  pageSummary: string;
}