# lib/api-zod/src/generated/types/violation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DesignGuard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FixAction } from "./fixAction";
import type { ViolationCategory } from "./violationCategory";
import type { ViolationSeverity } from "./violationSeverity";

export interface Violation {
  violationId: string;
  category: ViolationCategory;
  severity: ViolationSeverity;
  elementDescription: string;
  wcagCriterion: string;
  currentValue?: string;
  requiredValue?: string;
  fixDescription: string;
  fixAction?: FixAction;
}